Fix missionTool
Now it's fully functional. If weird things are happening you need to
normalize the line endings in all the files you are comparing.

TODOs:
- Automatically normalize line endings: Copy files to a temp folder, do
all comparisons in the temp folder, delete temp folder, copy files from
their original locations to the new mission location. Temp folder
location could be in the new mission folder. Prompt to use the
normalized files instead of the non-normalized ones if there is a
difference?
- Add a way to have warning outputs sent to a file instead of/as well as
the console
- Add a way to show a diff between manual merge files, or output it to a
file at least?
- Add a way to update multiple missions at once.